Add some core components

+++ api/src/database/migrations/20221130195223-create-type.js
@@ -27,7 +27,7 @@ module.exports = {
         defaultValue: true,
       },
       numberOfTopics: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.INTEGER,
         allowNull: true,
         defaultValue: 0,
       },
